+++ src/test/test_compiler.sh
@@ -75,6 +75,7 @@ TEST_FILES=(
     "src/test-resources/test_register.c"
     "src/test-resources/test_array_init.c"
     "src/test-resources/test_funcptr.c"
+    "src/test-resources/test_variadic.c"
 )
 
 mkdir -p build/test